fix/fix-move-component-correct-folder
