+++ src/stories/components/svg/ArrowUp.tsx
