+++ src/stories/components/table/TableCore.scss
